SENDER_TIMEOUT should be larger than 3 seconds
SENDER_TIMEOUT kicks in when the recipient URL is loaded. Within the
timeout, the recipient is supposed to talk to the mailbox server and
get the offer message with filename and file size. If this doesn't
happen, the BadCodeError is displayed. However, in some networks, it
may take more than 3 seconds to load. Increasing it here to 10s.