SENDER_TIMEOUT should be larger than 3 seconds (#80)
* SENDER_TIMEOUT should be larger than 3 seconds

SENDER_TIMEOUT kicks in when the recipient URL is loaded. Within the
timeout, the recipient is supposed to talk to the mailbox server and
get the offer message with filename and file size. If this doesn't
happen, the BadCodeError is displayed. However, in some networks, it
may take more than 3 seconds to load. Increasing it here to 10s.

* changing the connection timeout to 10 minutes

Mailbox (after both the parties are connected) is kept alive for 10
minutes, so it makes sense for this timeout to be the same.

* 10 x 60 instead of 60 x 10 just to be more clear and pedantic